% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonNormalize.R
\name{jsonNormalize}
\alias{jsonNormalize}
\title{Normalize a JSON string}
\usage{
jsonNormalize(jstring, prettify = FALSE)
}
\arguments{
\item{jstring}{a character string, the JSON string to be normalized, or
the path to a JSON file}

\item{prettify}{Boolean, whether to prettify the normalized JSON string}
}
\value{
The normalized JSON string.
}
\description{
Attempts to normalize or fix a JSON string.
}
\examples{
library(jsonNormalize)
# the keys of the following JSON string are not quoted
jstring <- "[{area:30,ind:[5,4.1,3.7],cluster:true},{ind:[],cluster:false}]"
cat(jsonNormalize(jstring, prettify = TRUE))
}
